First checkin of LillyDAP, the ever-so-simple LDAP library. This library aims to support lock-free concurrent operation. It basically translates BER byte streams to operations, and back. Operations have a parsed msgid, opcode and a struct-overlay with args. Dependencies on memory (region-based) and event loop are externalised.
Current functionality is close to working for the lillyget_xxx() side.
Testing has halted due to an issue that had to be filed against Quick DER,
https://github.com/vanrein/quick-der/issues/16
Aside from that and a few other small updates that were required, the
hope and expectation that Quick DER should hold up is indeed affirmative :-D

Who knows, some day... we will build this into Nginx, and have our own config
for LDAP proxy'ing in there...

ldap {
   server 123.45.67.89:389;

   location arpa2.org {
     ...
   };

   location @proxy {
      proxy_pass 11.22.33.44:389;
   };
}

Well, that's the stuff dreams are made of, I suppose?  Bits.  Both 1 and 0.